+++ src/recommendation/recommendation.module.ts
@@ -1,13 +1,11 @@
 import { Module } from "@nestjs/common";
-import { KeywordController } from "./keyword/keyword.controller";
-import { KeywordService } from "./keyword/keyword.service";
 import { DataModule } from "src/data/data.module";
-import { RegionController } from "./region/region.controller";
-import { RegionService } from "./region/region.service";
+import { RecommendationController } from "./recommendation.controller";
+import { RecommendationService } from "./recommendation.service";
 
 @Module({
    imports: [DataModule],
-   controllers: [KeywordController, RegionController],
-   providers: [KeywordService, RegionService],
+   controllers: [RecommendationController],
+   providers: [RecommendationService],
 })
 export class RecommendationModule {}
